requirements.txt
asgiref==3.7.2
Django==3.2.12
django-ckeditor==6.7.0
django-js-asset==2.1.0
django-mptt==0.14.0
persiantools==2.2.0
Pillow==10.0.1
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
